feat: packaging (#1)
* remove scripts for creating an asset package

* replace asset with package

This allows better reuse, samples and tests.
Furthermore, the packaging of `ros2-for-unity.xml` is now being
done by Unity as well as setting up the path and doing integrity checks.

* update README

* add samples

The old examples where moved into the new ones.

* add missing .meta files

* rename Scripts to Runtime

It seems that this is closer to the recommended layout

* rename pkg

Signed-off-by: Autumn60 <harada.akiro@gmail.com>

---------

Signed-off-by: Autumn60 <harada.akiro@gmail.com>
Co-authored-by: Eric Wolf <ewolf@advitec.de>